# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Financial Anomaly Detection API",
    "description": "Backend foundation for Machine Learning Enhanced Graph-Based Anomaly Detection for Financial Transactions. Provides transaction ingestion, graph construction, and behavioral/temporal feature extraction.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/anomalies/{dataset_id}/train": {
      "post": {
        "tags": [
          "Anomaly Detection (Machine Learning)"
        ],
        "summary": "Train Isolation Forest Model",
        "description": "Trains an unsupervised Isolation Forest model on the user-level feature matrix associated with the given dataset_id. Supports specifying an experiment_label, hyperparameter overrides (n_estimators, contamination, max_samples, random_state), and feature groups ('graph', 'behavioral', 'temporal').",
        "operationId": "train_anomaly_model_api_v1_anomalies__dataset_id__train_post",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          },
          {
            "name": "experiment_label",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional experiment label (defaults to payload.experiment_label or 'default')",
              "title": "Experiment Label"
            },
            "description": "Optional experiment label (defaults to payload.experiment_label or 'default')"
          },
          {
            "name": "split_label",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional split label for held-out training partition (overrides payload.split_label)",
              "title": "Split Label"
            },
            "description": "Optional split label for held-out training partition (overrides payload.split_label)"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AnomalyTrainRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "description": "Optional hyperparameter, feature group, and experiment label overrides.",
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnomalyTrainResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/anomalies/{dataset_id}/baseline": {
      "post": {
        "tags": [
          "Anomaly Detection (Machine Learning)"
        ],
        "summary": "Train Statistical Baseline (Non-ML)",
        "description": "Fits a non-ML rule-based z-score threshold statistical baseline. Flags users exceeding mean + z_threshold * std_dev across structural and behavioral features. Saved under experiment_label (default 'baseline_statistical') for uniform evaluation.",
        "operationId": "train_statistical_baseline_api_v1_anomalies__dataset_id__baseline_post",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/BaselineTrainRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "description": "Configuration for z-score threshold rule.",
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnomalyTrainResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/anomalies/{dataset_id}/experiments": {
      "get": {
        "tags": [
          "Anomaly Detection (Machine Learning)"
        ],
        "summary": "List All Trained Experiments for Dataset",
        "description": "Lists all experiment_labels currently trained for a dataset, along with feature groups and hyperparameters.",
        "operationId": "list_dataset_experiments_api_v1_anomalies__dataset_id__experiments_get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/anomalies/{dataset_id}/users": {
      "get": {
        "tags": [
          "Anomaly Detection (Machine Learning)"
        ],
        "summary": "Get User Anomaly Scores and Classifications",
        "description": "Returns per-user anomaly detection results for a given experiment_label including raw continuous scores, relative 0-100 presentation risk scores (inverted percentile rank), suspicious/normal status, and explainability reason codes based on training population deviations.",
        "operationId": "get_user_anomalies_api_v1_anomalies__dataset_id__users_get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          },
          {
            "name": "experiment_label",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Experiment identifier (e.g. 'default', 'e1', 'baseline_statistical')",
              "default": "default",
              "title": "Experiment Label"
            },
            "description": "Experiment identifier (e.g. 'default', 'e1', 'baseline_statistical')"
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional user ID filter for single entity inspection",
              "title": "User Id"
            },
            "description": "Optional user ID filter for single entity inspection"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Page size limit",
              "default": 50,
              "title": "Limit"
            },
            "description": "Page size limit"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Page offset index",
              "default": 0,
              "title": "Offset"
            },
            "description": "Page offset index"
          },
          {
            "name": "suspicious_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "If true, returns only entities flagged as suspicious",
              "default": false,
              "title": "Suspicious Only"
            },
            "description": "If true, returns only entities flagged as suspicious"
          },
          {
            "name": "split_label",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional split label to evaluate held-out test partition users",
              "title": "Split Label"
            },
            "description": "Optional split label to evaluate held-out test partition users"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAnomalyListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/anomalies/{dataset_id}/summary": {
      "get": {
        "tags": [
          "Anomaly Detection (Machine Learning)"
        ],
        "summary": "Get Dataset Anomaly Summary",
        "description": "Returns aggregate anomaly statistics (total entities, suspicious count, normal count, contamination rate, feature groups) along with the active model's metadata for the specified experiment.",
        "operationId": "get_anomaly_summary_api_v1_anomalies__dataset_id__summary_get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          },
          {
            "name": "experiment_label",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Experiment identifier (e.g. 'default', 'e1', 'baseline_statistical')",
              "default": "default",
              "title": "Experiment Label"
            },
            "description": "Experiment identifier (e.g. 'default', 'e1', 'baseline_statistical')"
          },
          {
            "name": "split_label",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional split label to compute summary over test partition",
              "title": "Split Label"
            },
            "description": "Optional split label to compute summary over test partition"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnomalySummaryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/anomalies/{dataset_id}/model": {
      "get": {
        "tags": [
          "Anomaly Detection (Machine Learning)"
        ],
        "summary": "Get Trained Model Metadata",
        "description": "Returns the full stored model metadata (algorithm, sklearn version, hyperparameters, exact feature names and groups used, entity count, training timestamp) for the specified experiment.",
        "operationId": "get_model_metadata_api_v1_anomalies__dataset_id__model_get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          },
          {
            "name": "experiment_label",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Experiment identifier (e.g. 'default', 'e1', 'baseline_statistical')",
              "default": "default",
              "title": "Experiment Label"
            },
            "description": "Experiment identifier (e.g. 'default', 'e1', 'baseline_statistical')"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelMetadata"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnomalySummaryResponse": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id",
            "description": "Dataset identifier"
          },
          "experiment_label": {
            "type": "string",
            "title": "Experiment Label",
            "description": "Experiment identifier",
            "default": "default"
          },
          "total_entities": {
            "type": "integer",
            "title": "Total Entities",
            "description": "Total number of evaluated entities"
          },
          "suspicious_count": {
            "type": "integer",
            "title": "Suspicious Count",
            "description": "Count of entities flagged as suspicious"
          },
          "normal_count": {
            "type": "integer",
            "title": "Normal Count",
            "description": "Count of entities classified as normal"
          },
          "contamination_rate": {
            "type": "number",
            "title": "Contamination Rate",
            "description": "Proportion of entities classified as suspicious"
          },
          "feature_groups": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Feature Groups",
            "description": "Feature groups used during training"
          },
          "feature_count": {
            "type": "integer",
            "title": "Feature Count",
            "description": "Number of features used during training"
          },
          "model_metadata": {
            "$ref": "#/components/schemas/ModelMetadata",
            "description": "Complete stored model metadata"
          }
        },
        "type": "object",
        "required": [
          "dataset_id",
          "total_entities",
          "suspicious_count",
          "normal_count",
          "contamination_rate",
          "feature_groups",
          "feature_count",
          "model_metadata"
        ],
        "title": "AnomalySummaryResponse",
        "description": "High-level summary of anomaly distribution and model configuration."
      },
      "AnomalyTrainRequest": {
        "properties": {
          "experiment_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experiment Label",
            "description": "Experiment identifier (e.g. 'default', 'e1', 'e2', 'baseline_statistical').",
            "default": "default"
          },
          "feature_groups": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feature Groups",
            "description": "Feature group subsets to include in training ('graph', 'behavioral', 'temporal').",
            "default": [
              "graph",
              "behavioral",
              "temporal"
            ]
          },
          "n_estimators": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "N Estimators",
            "description": "Number of base isolation trees in the ensemble.",
            "default": 100
          },
          "contamination": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 0.5,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Contamination",
            "description": "Expected proportion of outliers in the data set.",
            "default": 0.1
          },
          "max_samples": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Samples",
            "description": "Number of samples to draw from features to train each tree ('auto', int, or float (0, 1]).",
            "default": "auto"
          },
          "random_state": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Random State",
            "description": "Deterministic random seed for reproducible tree splits.",
            "default": 42
          },
          "split_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Split Label",
            "description": "Optional dataset split label for held-out training partition. If omitted, trains on full population (in_sample)."
          }
        },
        "type": "object",
        "title": "AnomalyTrainRequest",
        "description": "Configuration payload for training an Isolation Forest model on a dataset."
      },
      "AnomalyTrainResponse": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "success"
          },
          "model_metadata": {
            "$ref": "#/components/schemas/ModelMetadata"
          }
        },
        "type": "object",
        "required": [
          "dataset_id",
          "message",
          "model_metadata"
        ],
        "title": "AnomalyTrainResponse",
        "description": "Response returned upon successful model training."
      },
      "BaselineTrainRequest": {
        "properties": {
          "experiment_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experiment Label",
            "description": "Experiment identifier for statistical baseline.",
            "default": "baseline_statistical"
          },
          "z_threshold": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Z Threshold",
            "description": "Z-score threshold for classifying an entity as anomalous (default: 2.0, representing ~2.28% upper tail).",
            "default": 2.0
          },
          "features": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Features",
            "description": "List of graph feature names to evaluate using population z-scores (default: weighted_out_degree, betweenness_centrality, total_degree).",
            "default": [
              "weighted_out_degree",
              "betweenness_centrality",
              "total_degree"
            ]
          },
          "split_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Split Label",
            "description": "Optional dataset split label for held-out baseline fitting."
          }
        },
        "type": "object",
        "title": "BaselineTrainRequest",
        "description": "Configuration payload for training a non-ML rule-based statistical baseline."
      },
...
      "ExperimentListResponse": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id",
            "description": "Dataset identifier"
          },
          "total_experiments": {
            "type": "integer",
            "title": "Total Experiments",
            "description": "Total count of trained experiments stored for this dataset"
          },
          "experiments": {
            "items": {
              "$ref": "#/components/schemas/ModelMetadata"
            },
            "type": "array",
            "title": "Experiments",
            "description": "List of stored model metadata objects"
          }
        },
        "type": "object",
        "required": [
          "dataset_id",
          "total_experiments",
          "experiments"
        ],
        "title": "ExperimentListResponse",
        "description": "List of all trained experiments currently stored for a dataset."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UserAnomalyListResponse": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id",
            "description": "Dataset identifier"
          },
          "experiment_label": {
            "type": "string",
            "title": "Experiment Label",
            "description": "Experiment identifier",
            "default": "default"
          },
          "total_users": {
            "type": "integer",
            "title": "Total Users",
            "description": "Total users evaluated"
          },
          "suspicious_count": {
            "type": "integer",
            "title": "Suspicious Count",
            "description": "Total users flagged as suspicious (prediction == -1)"
          },
          "normal_count": {
            "type": "integer",
            "title": "Normal Count",
            "description": "Total users classified as normal (prediction == 1)"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "description": "Page size limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset",
            "description": "Page offset"
          },
          "users": {
            "items": {
              "$ref": "#/components/schemas/UserAnomalyResult"
            },
            "type": "array",
            "title": "Users",
            "description": "List of user anomaly scores and classifications"
          }
        },
        "type": "object",
        "required": [
          "dataset_id",
          "total_users",
          "suspicious_count",
          "normal_count",
          "limit",
          "offset",
          "users"
        ],
        "title": "UserAnomalyListResponse",
        "description": "Paginated collection of per-user anomaly detection results."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}